+++ CMakeLists.txt
@@ -240,11 +240,7 @@ if(TNN_HUAWEI_NPU_ENABLE)
         )
     endif()
     add_subdirectory(source/tnn/device/huawei_npu)
-    if(TNN_CPU_ENABLE)
-        set(TARGET_OBJECTS ${TARGET_OBJECTS} "$<TARGET_OBJECTS:TNNNPU>")
-    else()
-        set(TARGET_OBJECTS ${TARGET_OBJECTS} "$<TARGET_OBJECTS:TNNNPU>" "$<TARGET_OBJECTS:TNNBLOB>")
-    endif()
+    set(TARGET_OBJECTS ${TARGET_OBJECTS} "$<TARGET_OBJECTS:TNNNPU>")
 endif()
 
 if(TNN_RK_NPU_ENABLE)
